database/requirements.txt
pika
mysql-connector-python